Refactor document management components by removing EmployeeFileCard and consolidating file card functionality into a new FileCard component
